v8.2.4: Process every eligible patient release row
